+++ datacontract/breaking/breaking.py
@@ -1,6 +1,218 @@
 from datacontract.breaking.breaking_rules import BreakingRules
 from datacontract.model.breaking_change import BreakingChange, Location, Severity
-from datacontract.model.data_contract_specification import Field, Model, Quality
+from datacontract.model.data_contract_specification import Contact, Field, Info, Model, Quality, Terms
+
+
+def info_breaking_changes(
+    old_info: Info,
+    new_info: Info,
+    new_path: str,
+    include_severities: [Severity],
+) -> list[BreakingChange]:
+    results = list[BreakingChange]()
+
+    composition = ["info"]
+
+    if old_info and new_info:
+        info_definition_fields = vars(new_info) | new_info.model_extra | old_info.model_extra
+
+        for info_definition_field in info_definition_fields.keys():
+            if info_definition_field == "contact":
+                continue
+
+            old_value = getattr(old_info, info_definition_field, None)
+            new_value = getattr(new_info, info_definition_field, None)
+
+            rule_name = None
+            description = None
+
+            if old_value is None and new_value is not None:
+                rule_name = f"info_{_camel_to_snake(info_definition_field)}_added"
+                description = f"added with value: `{new_value}`"
+
+            elif old_value is not None and new_value is None:
+                rule_name = f"info_{_camel_to_snake(info_definition_field)}_removed"
+                description = "removed info property"
+
+            elif old_value != new_value:
+                rule_name = f"info_{_camel_to_snake(info_definition_field)}_updated"
+                description = f"changed from `{old_value}` to `{new_value}`"
+
+            if rule_name is not None:
+                severity = _get_rule(rule_name)
+                if severity in include_severities:
+                    results.append(
+                        BreakingChange(
+                            description=description,
+                            check_name=rule_name,
+                            severity=severity,
+                            location=Location(path=new_path, composition=composition + [info_definition_field]),
+                        )
+                    )
+
+        results.extend(
+            contact_breaking_changes(
+                old_contact=getattr(old_info, "contact", None),
+                new_contact=getattr(new_info, "contact", None),
+                composition=composition + ["contact"],
+                new_path=new_path,
+                include_severities=include_severities,
+            )
+        )
+
+    return results
+
+
+def contact_breaking_changes(
+    old_contact: Contact,
+    new_contact: Contact,
+    composition: list[str],
+    new_path: str,
+    include_severities: [Severity],
+) -> list[BreakingChange]:
+    results = list[BreakingChange]()
+
+    if not old_contact and new_contact:
+        rule_name = "contact_added"
+        severity = _get_rule(rule_name)
+        description = "added contact"
+
+        if severity in include_severities:
+            results.append(
+                BreakingChange(
+                    description=description,
+                    check_name=rule_name,
+                    severity=severity,
+                    location=Location(path=new_path, composition=composition),
+                )
+            )
+
+    elif old_contact and not new_contact:
+        rule_name = "contact_removed"
+        severity = _get_rule(rule_name)
+        description = "removed contact"
+
+        if severity in include_severities:
+            results.append(
+                BreakingChange(
+                    description=description,
+                    check_name=rule_name,
+                    severity=severity,
+                    location=Location(path=new_path, composition=composition),
+                )
+            )
+
+    elif old_contact and new_contact:
+        contact_definition_fields = vars(new_contact) | new_contact.model_extra | old_contact.model_extra
+
+        for contact_definition_field in contact_definition_fields.keys():
+            old_value = getattr(old_contact, contact_definition_field, None)
+            new_value = getattr(new_contact, contact_definition_field, None)
+
+            rule_name = None
+            description = None
+
+            if old_value is None and new_value is not None:
+                rule_name = f"contact_{_camel_to_snake(contact_definition_field)}_added"
+                description = f"added with value: `{new_value}`"
+
+            elif old_value is not None and new_value is None:
+                rule_name = f"contact_{_camel_to_snake(contact_definition_field)}_removed"
+                description = "removed contact property"
+
+            elif old_value != new_value:
+                rule_name = f"contact_{_camel_to_snake(contact_definition_field)}_updated"
+                description = f"changed from `{old_value}` to `{new_value}`"
+
+            if rule_name is not None:
+                severity = _get_rule(rule_name)
+                if severity in include_severities:
+                    results.append(
+                        BreakingChange(
+                            description=description,
+                            check_name=rule_name,
+                            severity=severity,
+                            location=Location(path=new_path, composition=composition + [contact_definition_field]),
+                        )
+                    )
+
+    return results
+
+
+def terms_breaking_changes(
+    old_terms: Terms,
+    new_terms: Terms,
+    new_path: str,
+    include_severities: [Severity],
+) -> list[BreakingChange]:
+    results = list[BreakingChange]()
+
+    composition = ["terms"]
+
+    if not old_terms and new_terms:
+        rule_name = "terms_added"
+        severity = _get_rule(rule_name)
+        description = "added terms"
+
+        if severity in include_severities:
+            results.append(
+                BreakingChange(
+                    description=description,
+                    check_name=rule_name,
+                    severity=severity,
+                    location=Location(path=new_path, composition=composition),
+                )
+            )
+    elif old_terms and not new_terms:
+        rule_name = "terms_removed"
+        severity = _get_rule(rule_name)
+        description = "removed terms"
+
+        if severity in include_severities:
+            results.append(
+                BreakingChange(
+                    description=description,
+                    check_name=rule_name,
+                    severity=severity,
+                    location=Location(path=new_path, composition=composition),
+                )
+            )
+
+    if old_terms and new_terms:
+        terms_definition_fields = vars(new_terms) | new_terms.model_extra | old_terms.model_extra
+
+        for terms_definition_field in terms_definition_fields.keys():
+            old_value = getattr(old_terms, terms_definition_field, None)
+            new_value = getattr(new_terms, terms_definition_field, None)
+
+            rule_name = None
+            description = None
+
+            if old_value is None and new_value is not None:
+                rule_name = f"terms_{_camel_to_snake(terms_definition_field)}_added"
+                description = f"added with value: `{new_value}`"
+
+            elif old_value is not None and new_value is None:
+                rule_name = f"terms_{_camel_to_snake(terms_definition_field)}_removed"
+                description = "removed info property"
+
+            elif old_value != new_value:
+                rule_name = f"terms_{_camel_to_snake(terms_definition_field)}_updated"
+                description = f"changed from `{old_value}` to `{new_value}`"
+
+            if rule_name is not None:
+                severity = _get_rule(rule_name)
+                if severity in include_severities:
+                    results.append(
+                        BreakingChange(
+                            description=description,
+                            check_name=rule_name,
+                            severity=severity,
+                            location=Location(path=new_path, composition=composition + [terms_definition_field]),
+                        )
+                    )
+
+    return results
 
 
 def quality_breaking_changes(
@@ -129,14 +341,14 @@ def model_breaking_changes(
 ) -> list[BreakingChange]:
     results = list[BreakingChange]()
 
-    model_definition_fields = vars(new_model)
+    model_definition_fields = vars(new_model) | new_model.model_extra | old_model.model_extra
 
     for model_definition_field in model_definition_fields.keys():
         if model_definition_field == "fields":
             continue
 
-        old_value = getattr(old_model, model_definition_field)
-        new_value = getattr(new_model, model_definition_field)
+        old_value = getattr(old_model, model_definition_field, None)
+        new_value = getattr(new_model, model_definition_field, None)
 
         rule_name = None
         description = None
@@ -237,13 +449,13 @@ def field_breaking_changes(
 ) -> list[BreakingChange]:
     results = list[BreakingChange]()
 
-    field_definition_fields = vars(new_field)
+    field_definition_fields = vars(new_field) | new_field.model_extra | old_field.model_extra
     for field_definition_field in field_definition_fields.keys():
         if field_definition_field == "ref_obj":
             continue
 
-        old_value = getattr(old_field, field_definition_field)
-        new_value = getattr(new_field, field_definition_field)
+        old_value = getattr(old_field, field_definition_field, None)
+        new_value = getattr(new_field, field_definition_field, None)
 
         if field_definition_field == "fields":
             results.extend(
@@ -320,9 +532,15 @@ def _get_rule(rule_name) -> Severity:
     try:
         return getattr(BreakingRules, rule_name)
     except AttributeError:
-        print(f"WARNING: Breaking Rule not found for {rule_name}!")
-        return Severity.ERROR
+        try:
+            first, *_, last = rule_name.split("_")
+            short_rule = "__".join([first, last])
+            return getattr(BreakingRules, short_rule)
+        except AttributeError:
+            print(f"WARNING: Breaking Rule not found for {rule_name}!")
+            return Severity.ERROR
 
 
 def _camel_to_snake(s):
+    s = s.replace("-", "_")
     return "".join(["_" + c.lower() if c.isupper() else c for c in s]).lstrip("_")