+++ CHANGELOG.md
@@ -14,9 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `suite_name` the name of the expectation suite to export
   - `engine` used to run checks
   - `sql_server_type` to define the type of SQL Server to use when engine is `sql`
-
+- Changelog support for `Info` and `Terms` blocks.
 
 ### Changed
+- Changelog support for custom extension keys in `Models` and `Fields` blocks.
 
 ### Fixed
 - raise valid exception in DataContractSpecification.from_file if file does not exist